+++ packages/base-controller/src/RestrictedMessenger.ts
@@ -127,16 +127,17 @@ export class RestrictedMessenger<
   }
 
   /**
-   * Registers action handlers for a list of methods on a class instance
+   * Registers action handlers for a list of methods on a messenger client
    *
-   * @param instance - The class instance with a name property and methods
-   * @param methodNames - The names of the methods to register as action handlers
+   * @param messengerClient - The object that is expected to make use of the messenger.
+   * @param methodNames - The names of the methods on the messenger client to register as action
+   * handlers
    */
   registerMethodActionHandlers<
-    Instance extends { name: string },
-    MethodNames extends keyof Instance & string,
-  >(instance: Instance, methodNames: readonly MethodNames[]) {
-    this.#messenger.registerMethodActionHandlers(instance, methodNames);
+    MessengerClient extends { name: string },
+    MethodNames extends keyof MessengerClient & string,
+  >(messengerClient: MessengerClient, methodNames: readonly MethodNames[]) {
+    this.#messenger.registerMethodActionHandlers(messengerClient, methodNames);
   }
 
   /**