+++ packages/base-controller/CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `registerMethodActionHandlers` method to `Messenger`, and `RestrictedMessenger` for simplified bulk action handler registration ([#5927](https://github.com/MetaMask/core/pull/5927))
-  - Allows registering multiple action handlers at once by passing an array of method names
-  - Automatically binds methods to the controller instance
+  - Allows registering action handlers that map to methods on a messenger client at once by passing an array of method names
+  - Automatically binds action handlers to the given messenger client
 
 ### Changed